refactor(mini-player): extract the swipe rail into a hook
Move the track-skip swipe out of MiniPlayer into useSwipeRail so the
component is back under the giant-component threshold and changes for one
reason. Flush an interrupted commit's deferred skip when a fresh press
lands mid-slide, so a rapid re-grab no longer strands the pushed-off
track. Share the cover thumbnail via TrackArtwork.
